Title: 낮은 지점들, Time: 66ms, Memory: 2MB, Status: Wrong Answer - Codetree

+++ 250606/낮은 지점들/lowest-points.cpp
@@ -7,7 +7,7 @@ int n;
 int x[100000], y[100000];
 
 int main() {
-    unordered_map<int,int> point;
+    unordered_map<long long,long long> point;
 
     cin >> n;
     for (int i = 0; i < n; i++) {
